disease_surveillance/requirements.txt
# Disease Hotspot Detection System Requirements

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Time Series Analysis
statsmodels>=0.14.0

# Spatial Analysis
scikit-learn>=1.3.0
geopy>=2.3.0

# Data Ingestion
openpyxl>=3.1.0
xlrd>=2.0.1

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Hints
typing-extensions>=4.7.0
